Repo Titles: Refactor to use RepoIdToString
Currently we rely on two separate components, <PrototypesPage />
and <ProjectDetail />, to render the logic of converting a repoId
to a project title. Since the logic is the same in both components,
 and assuming that we want to maintain that consistency in the future,
it would help minimize errors to capture that logic in a function to use
 in both components. We also already have a function that encapsulates
that logic already, repoIdtoString.

The expressiveness of repoIdToString in this context probably could
be improved?

Test Plan:

Logic to display the project title is already
tested in the <App ./> component in App.js.

The second area we need to test is the logic in PrototypePage.js