Add graph node count to PagerankTable Connected to #576 [1/2]
Adds a label that lists the number of nodes in the graph. The label appear
after the graph has been loaded, above the Node Filter, and uses `pagerankNodeDecomposition.size`
to calculate the node count.

Test Plan:

Added a unit test that verifies the parent container, PagerankTable,
contains a <p> tag with the correct title, ie lists the node count
that matches `matches pagerankNodeDecomposition.size`

Visual tests verify the location of the node count label above the
Node Filter,

Screenshot attached to Github description: