[WIP] Add optional node filtering to pagerankGraph
In line with the notion of pagerankGraph as a wrapper around graph,
adding optional prefix filters to nodes amounts to adding an optional prefix
parameter to `nodes()` and passing that prefix to `_graph`, which
performs the filtering, and returning that nodes provides by `_graph`
post-filter and augmenting each result with a score.

Test Plan:
Yarn Test passes
@decentraltion provided some guidance in the Discord on what to test -
to test that the parameter was passed down correctly, the unit tests
compare the nodes returned in `graph` and `pagerankGraph` after filtering
each using the same options. I used the same test cases used in `Graph.test`
for filtering:
1) "uses empty prefix when no options object"
2) "requires a prefix when options are specified"
3) "does a prefix filter"
4) "yields nothing when prefix matches nothing"